+++ src/test/java/cart/domain/user/EmailTest.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.user;
 
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
